Up to ep 39

+++ Assets/Project/Scripts/UI/StaminaBar.cs
@@ -9,12 +9,12 @@ namespace PHH
     {
         public Slider slider;
 
-        public void SetCurrentStamina(int currentStamina)
+        public void SetCurrentStamina(float currentStamina)
         {
             slider.value = currentStamina;
         }
 
-        public void SetMaxStamina(int maxStamina)
+        public void SetMaxStamina(float maxStamina)
         {
             slider.maxValue = maxStamina;
             slider.value = maxStamina;
